FIX: Warnings

+++ LiverTumorSegmentation/LiverTumorSegmentation.cxx
@@ -428,13 +428,13 @@ void LiverTumorSegmentation::OnSegmentationParametersOk( int module )
      case CONFIDENCE_CONNECTED:
       this->m_ConfidenceConnectedModule.SetMultiplier( atof(m_Parameter001->value() ) );
       this->m_ConfidenceConnectedModule.SetNumberOfIterations( atof(m_Parameter002->value() ) );
-      this->m_ConfidenceConnectedModule.SetInitialNeighborhoodRadius( atof(m_Parameter003->value() ) );
+      this->m_ConfidenceConnectedModule.SetInitialNeighborhoodRadius( (unsigned int)atof(m_Parameter003->value() ) );
       this->m_ConfidenceConnectedModule.SetSmoothingTimeStep( atof(m_Parameter101->value() ) );
       this->m_ConfidenceConnectedModule.SetSmoothingIterations( atoi(m_Parameter102->value() ) );
       break;
     case CONNECTED_THRESHOLD:
-      this->m_ConnectedThresholdModule.SetLowerThreshold( atof(m_Parameter001->value() ) );
-      this->m_ConnectedThresholdModule.SetUpperThreshold( atof(m_Parameter002->value() ) );
+      this->m_ConnectedThresholdModule.SetLowerThreshold( (unsigned int)atof(m_Parameter001->value() ) );
+      this->m_ConnectedThresholdModule.SetUpperThreshold( (unsigned int)atof(m_Parameter002->value() ) );
       this->m_ConnectedThresholdModule.SetSmoothingTimeStep( atof(m_Parameter101->value() ) );
       this->m_ConnectedThresholdModule.SetSmoothingIterations( atoi(m_Parameter102->value() ) );
       break;
